Reverting a change from commit 52ac3036aea3cd1d89f1fc8b4102ed90d262f3f1
We deactivated an optimization in Component's render implementation,
which with the makes sense, even not being optimal.

The bug claimed by #311 is not a bug in Component but a violation
of the 'data-down' approach taken with our implementation.
Tl;tr Never derive child props from parent state dynamically.
Instead provide a set of derived props to the component(here ContextSection).